% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATools.R
\name{get_point_dist}
\alias{get_point_dist}
\title{-------------------------------------------------
CA tools}
\usage{
get_point_dist(.tbl)
}
\description{
対応分析に関係する小道具集
}
\details{
前提としているCAは、package FactoMineR のCA
2015−10−23
kazuo.fujimoto2007@gmail.com

重心-プロファイルポイント間距離
.tbl 度数（クロス）表
}
