% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl2_1.R
\docType{data}
\name{tbl2_1}
\alias{tbl2_1}
\title{捜査された犯罪と発生場所：表2_1}
\format{
A table data with 10 rows and 6 cols:
\describe{
  \item{行 地域：}{オスロ,中部地域,北部地域,合計,全国}
  \item{列 犯罪}{強盗,詐欺.破壊,合計.密,疎}
}
}
\source{
『対応分析入門』第２章より抜粋
}
\usage{
data(tbl2_1)
}
\description{
このデータセット（クロス表）は、『対応分析入門』第２章でCA（対応分析）の事例として
使われているものです。
}
\examples{
# データの表示
tbl2_1

}
\keyword{datasets}
